requirements.txt
Django==5.1.5
mysqlclient==2.2.4
python-dotenv==1.0.1
django-jazzmin==3.0.0
django-ckeditor==6.7.1
markdown==3.7
Pillow==11.0.0
Pygments==2.19.1
gunicorn==23.0.0
whitenoise==6.8.2